Renamed triangle1 -> triangle_ui.

+++ clean_all_samples.bat
@@ -4,8 +4,8 @@ cd samples
 for /D %%G in (*) do (
     cd %%G
     echo %%G
-    rmdir /s /q zig-out
-    rmdir /s /q zig-cache
+    if exist zig-out rmdir /s /q zig-out
+    if exist zig-cache rmdir /s /q zig-cache
     cd..
 )
 cd..
